requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
numpy==2.3.2
traci==1.26.0
sumolib==1.26.0
pytest==8.3.2
httpx==0.27.2
SQLAlchemy==2.0.32
alembic==1.13.2
authlib==1.3.1
python-jose[cryptography]==3.3.0
psycopg==3.2.1
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
